pyproject.toml
[project]
name = "structured-outputs-with-pydantic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.5.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.1",
]
